% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{email_send_single}
\alias{email_send_single}
\title{Send a single email through Postmark}
\usage{
email_send_single(
  from,
  to,
  msg_stream,
  subject = NULL,
  html_body = NULL,
  text_body = NULL,
  token = NULL
)
}
\arguments{
\item{from}{Character scalar. Email address of the sender.}

\item{to}{Character vector. Email addresses of recipients (max 50).}

\item{msg_stream}{Character scalar. Either "outbound" or "broadcast".}

\item{subject}{Character scalar. Email subject line.}

\item{html_body}{Character scalar. HTML content of the email.}

\item{text_body}{Character scalar. Plain text content of the email.}

\item{token}{Character string. Your Postmark API token. If NULL (default), the
function will attempt to retrieve the token using \code{get_token()}
\code{POSTMARK_SERVER_TOKEN} as an environment variable. Check the Postmark's API
documentation on how to get a server token.}
}
\value{
A list containing the response details.
}
\description{
This function sends a single email via the Postmark API service. It supports
both HTML and plain text email formats (but not both simultaneously) and can
be used for both transactional ("outbound") and bulk ("broadcast") email
message streams.
}
\examples{
\dontrun{
 email_send_single(
  from = "sender@example.com",
  to = "recipient@example.com",
  msg_stream = "outbound",
  subject = "Hello from R",
  html_body = "<h1>Hello world!</h1>",
)
}
}
