% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmarkr-package.R
\docType{package}
\name{postmarkr-package}
\alias{postmarkr}
\alias{postmarkr-package}
\title{postmarkr: Email Delivery Made Simple: Postmark API Integration for R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides functions to interact with the Postmark API from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andreranza.github.io/postmarkr/}
}

}
\author{
\strong{Maintainer}: Andrea Ranzato \email{andrea@nutrivetpet.com}

Other contributors:
\itemize{
  \item nutrivetpet Srls [funder]
}

}
\keyword{internal}
