% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{postmarkr_response}
\alias{postmarkr_response}
\title{Response from Postmark API}
\usage{
postmarkr_response(
  data = NULL,
  status = integer(0),
  request = NULL,
  response = NULL,
  success = logical(0)
)
}
\description{
Structured response object
}
