% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmarkr-package.R
\docType{package}
\name{postmarkr-package}
\alias{postmarkr}
\alias{postmarkr-package}
\title{postmarkr: Email Delivery Made Simple: Unofficial Postmark API Integration for R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions for interacting with the Postmark API from R. This package is not created by or affiliated with Postmark.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nutrivetpet.github.io/postmarkr}
  \item \url{https://github.com/nutrivetpet/postmarkr}
  \item Report bugs at \url{https://github.com/nutrivetpet/postmarkr/issues}
}

}
\author{
\strong{Maintainer}: Andrea Ranzato \email{andrea@nutrivetpet.com}

Other contributors:
\itemize{
  \item nutrivetpet Srls [funder, copyright holder]
}

}
\keyword{internal}
